Major refactor to clean up how everything works
