+++ expansionhub/CMakeLists.txt
@@ -4,7 +4,14 @@ set(CMAKE_EXE_LINKER_FLAGS "-Wl,--allow-shlib-undefined")
 
 add_executable(ExpansionHubDaemon
     main.cpp
-    serial.c
+    PidConstants.cpp
+    SerialPort.cpp
+    EnabledState.cpp
+    MotorNtState.cpp
+    ServoNtState.cpp
+    ExpansionHubNtState.cpp
+    ExpansionHubSerial.cpp
+    SystemDUsbMonitor.cpp
 )
 target_compile_features(ExpansionHubDaemon PRIVATE cxx_std_20)
 target_compile_features(ExpansionHubDaemon PRIVATE c_std_17)
