+++ expansionhub/SerialPort.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+
+namespace eh {
+
+int OpenRhspSerialPort(const char* serialPort);
+
+}
