--- expansionhub/UvSerial.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "wpinet/uv/Poll.h"
-#include <memory>
